refactoring app to segregate items
